flask/requirements.txt
astroid==2.3.3
Click==7.0
colorama==0.4.3
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
six==1.13.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
wfastcgi==3.0.0
wrapt==1.11.2
